Update python binding and CMake system. * CMake now properly creates 'antimony.py' from createAntimonyLib.py * CellML off by default (sorry, CellML) * libz included if libxml needed (since the latter uses the former) * Fix warnings getch->_getch and size_t. * Everything now installed for python that needs to be installed. * createAntimonyLib.py now takes the directory of the header file as an argument, so it can be called by CMake. * setup now uses distutils.core instead of setuptools, so it won't try to install as an egg file (which wouldn't work, now that we're reading the .dll directly).

+++ antimony/src/testantimony.cpp
@@ -28,7 +28,7 @@ int main(int argc, char** argv)
     }
   }
   for (size_t file=0; file<getNumFiles(); file++) {
-    revertTo(file+1);
+    revertTo(static_cast<long>(file)+1);
     size_t nummods = getNumModules();
     char** modnames = getModuleNames();
     string filename(argv[file+1]);
